remove old comment entity
