+++ packages/server/customer-os-api/dataloader/comment_dataloader.go
@@ -3,22 +3,22 @@ package dataloader
 import (
 	"context"
 	"github.com/graph-gophers/dataloader"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/entity"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/tracing"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/utils"
+	neo4jentity "github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/entity"
 	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
 	"github.com/pkg/errors"
 	"reflect"
 )
 
-func (i *Loaders) GetCommentsForIssue(ctx context.Context, issueId string) (*entity.CommentEntities, error) {
+func (i *Loaders) GetCommentsForIssue(ctx context.Context, issueId string) (*neo4jentity.CommentEntities, error) {
 	thunk := i.CommentsForIssue.Load(ctx, dataloader.StringKey(issueId))
 	result, err := thunk()
 	if err != nil {
 		return nil, err
 	}
-	resultObj := result.(entity.CommentEntities)
+	resultObj := result.(neo4jentity.CommentEntities)
 	return &resultObj, nil
 }
 
@@ -43,12 +43,12 @@ func (b *commentBatcher) getCommentsForIssues(ctx context.Context, keys dataload
 		return []*dataloader.Result{{Data: nil, Error: err}}
 	}
 
-	commentEntitiesByMeetingId := make(map[string]entity.CommentEntities)
+	commentEntitiesByMeetingId := make(map[string]neo4jentity.CommentEntities)
 	for _, val := range *commentEntitiesPtr {
 		if list, ok := commentEntitiesByMeetingId[val.DataloaderKey]; ok {
 			commentEntitiesByMeetingId[val.DataloaderKey] = append(list, val)
 		} else {
-			commentEntitiesByMeetingId[val.DataloaderKey] = entity.CommentEntities{val}
+			commentEntitiesByMeetingId[val.DataloaderKey] = neo4jentity.CommentEntities{val}
 		}
 	}
 
@@ -61,10 +61,10 @@ func (b *commentBatcher) getCommentsForIssues(ctx context.Context, keys dataload
 		}
 	}
 	for _, ix := range keyOrder {
-		results[ix] = &dataloader.Result{Data: entity.CommentEntities{}, Error: nil}
+		results[ix] = &dataloader.Result{Data: neo4jentity.CommentEntities{}, Error: nil}
 	}
 
-	if err = assertEntitiesType(results, reflect.TypeOf(entity.CommentEntities{})); err != nil {
+	if err = assertEntitiesType(results, reflect.TypeOf(neo4jentity.CommentEntities{})); err != nil {
 		tracing.TraceErr(span, err)
 		return []*dataloader.Result{{nil, err}}
 	}
